Rename ghcn website example to orthographic
