+++ examples/website/orthographic/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "deckgl-examples-ghcn",
+  "name": "deckgl-examples-orthographic",
   "version": "0.0.0",
   "private": true,
   "license": "MIT",
